Generate API models as Go structs automagically
I copied over a subset of the JSON schema defined in PokeAPI/api-data
and used the go-jsonschema tool to generate Go structs that are very
close to the desired outcome. The manual tweaks required to get to the
desired state are pretty small, so I'm going to spend time trying to
get a perfect generator working.

I'll make the manual tweaks to the structs in a separate commit so it's
obvious what has been changed.