Setup a paginator interface
When we call an endpoint without an ID or name, the API returns a list
of NamedApiResource objects, which each have a link to an object of that
resource. I'm envisioning we can paginate through these top-level
GET requests, for each page we get N links to natures, pokemon, stats
etc.

Then within each iteration of the paginator, we can spawn N goroutines
to concurrently request a resource object. Once all N goroutines have
finished their work, we return a response to the user, marking the end
of that iteration.

We can then paginate through a response until we reach the end.