List objects of a resource concurrently
If you don't specify an ID or Name when calling a resource endpoint, the
API will return a paginated list of URLs to individual instances of that
resource.

We can set-up a paginator with N page size - where each iteration of the
paginator will make one request to get the list of resource URLs, and
then N requests to fetch the actual resources.

This is a nicer interface for users of the SDK. Instead of expecting a
client to make follow-up requests to fetch the resources themselves
this change does that for them.

Users make a ListX request, get an iterator, then call Next() on that
iterator to receive pages.